fix: break circular dependency T-76 between ArrayDataFrame.ts and processDataFrame.ts
Extract field-type-guessing helpers (`guessFieldTypeFromNameAndValue`,
`getFieldTypeFromValue`, `guessFieldTypeFromValue`, `guessFieldTypeForField`)
and the NUMBER regex from `processDataFrame.ts` into a new `guessFieldType.ts`.
Update all internal `grafana-data` consumers and `index.ts` to import from
`./guessFieldType` directly. Move related tests into `guessFieldType.test.ts`.

Side effect: also breaks T-77 (`processDataFrame.ts` <-> `DataFrameJSON.ts`),
since `DataFrameJSON.ts` only imported `guessFieldTypeFromNameAndValue` from
`processDataFrame.ts`.

Part of circular dependency cleanup effort.
See cat /users/hugoh/.claude/hugoh-memories/trivial-cycles.json for the full TODO list.
