+++ packages/grafana-data/src/utils/csv.ts
@@ -4,7 +4,7 @@ import Papa, { type ParseConfig, type Parser, type ParseResult } from 'papaparse
 
 // Types
 import { MutableDataFrame } from '../dataframe/MutableDataFrame';
-import { guessFieldTypeFromValue } from '../dataframe/processDataFrame';
+import { guessFieldTypeFromValue } from '../dataframe/guessFieldType';
 import { getFieldDisplayName } from '../field/fieldState';
 import { type DataFrame, type Field, type FieldConfig, FieldType } from '../types/dataFrame';
 import { formattedValueToString } from '../valueFormats/valueFormats';
